requirements.txt
jupyter
streamlit
matplotlib
scipy
tensorflow
numpy
Pillow
idna
huggingface-cli
huggingface_hub[cli]
